requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
fonttools==4.38.0
gensim==4.2.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
mail-parser==3.15.0
matplotlib==3.6.1
nltk==3.7
# numpy==1.23.4
numpy==1.22.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.0
scipy==1.9.3
simplejson==3.17.6
six==1.16.0
sklearn==0.0
smart-open==6.2.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
